# static/swaggerAuth.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API Documentation",
    "description": "API for marketplace.",
    "version": "1.0.3"
  },
  "paths": {
...
    "/shops/shop": {
      "post": {
        "summary": "Create shop or update shop info",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Create shop or update shop info",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdateShopSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shop created/updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Shop created/updated successfully"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/shops/shop_photo": {
      "post": {
        "summary": "Upload Shop Photo",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Shop photo upload",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ShopPhotoUploadSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shop photo uploaded successfully",
            "content": {
              "application/json": {
                "example": {
                  "photo_shop": "https://api.dorechi.store/static/media/shops/filename.png"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ShopNotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "413": {
            "$ref": "#/components/responses/EntityTooLarge3"
          }
        }
      },
      "delete": {
        "summary": "Delete Shop Photo",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Shop photo deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Shop photo deleted successfully"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ShopNotFound"
          }
        }
      },
      "get": {
        "summary": "Get Shop Photo",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Receive  url for shop photo",
            "content": {
              "application/json": {
                "example": {
                  "photo_shop": "https://api.dorechi.store/static/media/shops/filename.png"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ShopNotFound"
          }
        }
      }
    },
    "/shops/shop_banner": {
      "post": {
        "summary": "Upload Shop Banner",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Upload Shop Banner",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ShopBannerUploadSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shop banner uploaded successfully",
            "content": {
              "application/json": {
                "example": {
                  "banner_shop": "https://api.dorechi.store/static/media/banner_shops/filename.png"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ShopNotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "413": {
            "$ref": "#/components/responses/EntityTooLarge5"
          }
        }
      },
      "delete": {
        "summary": "Delete shop banner",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Shop banner deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Shop banner deleted successfully"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ShopNotFound"
          }
        }
      },
      "get": {
        "summary": "Get Shop Banner",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Receive absolute url for shop banner",
            "content": {
              "application/json": {
                "example": {
                  "banner_shop": "https://api.dorechi.store/static/media/banner_shops/filename.png"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ShopNotFound"
          }
        }
      }
    },
...
    "/shops/shop_info": {
      "get": {
        "summary": "Get Shop Information",
        "tags": [
          "Shop"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Shop information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopInfo"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/ShopNotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "EntityTooLarge3": {
        "description": "Request entity too large",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "File size too large. Maximum file size is 3MB."
                }
              }
            }
          }
        }
      },
      "EntityTooLarge5": {
        "description": "Request entity too large",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "File size too large. Maximum file size is 5MB."
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Missing Authorization Header"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request data"
                }
              }
            }
          }
        }
      },
      "ShopNotFound": {
        "description": "Shop not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Shop not found"
                }
              }
            }
          }
        }
      },
      "UserNotFound": {
        "description": "User not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User not found"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "loc": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "msg": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "loc": [
                        "phone_number"
                      ],
                      "msg": "Value error, Invalid phone number format. Must start with +380 and have 9 digits.",
                      "type": "value_error"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "CreateUpdateShopSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "phone_number": {
            "type": "string",
            "example": "+380000000000"
          },
          "link": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "phone_number"
        ]
      },
...
      "ShopPhotoUploadSchema": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "image"
        ]
      },
      "ShopBannerUploadSchema": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "image"
        ]
      },
...
      "ShopInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "owner_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "photo_shop": {
            "type": "string",
            "format": "uri",
            "description": "URL to the shop"
          },
          "banner_shop": {
            "type": "string",
            "format": "uri",
            "description": "URL to the shop"
          },
          "phone_number": {
            "type": "string"
          },
          "link": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Bearer token"
      }
    }
  }
}